# pkg/server/openapi.yaml
openapi: 3.1.0
info:
  title: native-ops daemon API
  version: set-at-runtime
  summary: Plan, deploy and inspect one Incus host from CI, an agent or a dashboard.
  description: |
    The native-ops daemon runs on the host it manages. Every change to the host is a **job** with a
    record; one change runs at a time (a second gets `409 busy`).

    **Deploys** are the normal way to change the host: push a protected `deploy-*` tag in the
    configuration repository and call `POST /v1/deploy` with it. The daemon reads that commit from the
    git server itself; nothing is uploaded, and the tag push is the approval. `POST /v1/deploy/plan`
    shows what a tag would change first; `GET /v1/deploys` says what the host runs now.

    **Roles** are ordered: viewer < planner < deployer < admin. Each operation names the least role it
    needs in `x-native-ops-role`. A token with a **scope** (for a fleet manager) may only call the
    operations marked `x-native-ops-scoped: true`, and only for instances its scope allows.

    Some operations exist only when the daemon is started with the flag in `x-native-ops-enabled-by`;
    otherwise they answer 404.

    Errors are `{"error": "<message>", "code": "<code>"}`. A failure to read the host is a generic
    `502 host_unavailable` that never echoes internal text.

    The same operations are available to AI agents as MCP tools at `POST /mcp` (see docs/api.md).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: "{daemon}"
...
security:
  - bearer: []
...
paths:
...
  /mcp:
    post:
      operationId: mcp
      tags: [meta]
      summary: Model Context Protocol endpoint
      description: |
        A stateless MCP server (Streamable HTTP, JSON answers). It takes a bearer token only, never a
        session cookie. Its tools call the operations here with the caller's token, so a tool can do
        exactly what the token could; `tools/list` shows only the tools the token's role and scope can use.
        Tools: whoami, host_status, list_deploys, plan_deploy, deploy, list_jobs, get_job, list_plans,
        get_plan, list_instances, get_instance, prune_images.
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: One JSON-RPC 2.0 message (batches are refused).
              required: [jsonrpc, method]
              properties:
                jsonrpc: { const: "2.0" }
                id: { type: [string, integer] }
                method: { type: string, examples: [initialize, tools/list, tools/call, ping] }
                params: { type: object }
      responses:
        "200": { description: "The JSON-RPC response.", content: { application/json: {} } }
        "202": { description: "A notification was accepted; there is no answer." }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: "The request came from a browser page of another origin." }
    get:
      operationId: mcpStream
      tags: [meta]
      summary: Not offered (no event stream)
      x-native-ops-role: none
      security: []
      responses:
        "405": { description: "The server opens no event stream; use POST." }
    delete:
      operationId: mcpEndSession
      tags: [meta]
      summary: Not offered (no sessions)
      x-native-ops-role: none
      security: []
      responses:
        "405": { description: "The server keeps no sessions." }
...
  /api/login:
    post:
      operationId: login
      tags: [session]
      summary: Local sign-in (sets the session cookie)
      security: []
      x-native-ops-role: none
      x-native-ops-enabled-by: --enable-auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              additionalProperties: false
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200": { description: "Signed in." }
        "400": { description: "Local sign-in is not enabled (local_signin_off)." }
        "401": { description: "Wrong username or password (bad_login)." }
        "429": { description: "Too many failed sign-ins for this account (too_many_attempts)." }
...
  /auth/oidc/callback:
    get:
      operationId: oidcCallback
      tags: [session]
      summary: OIDC callback
      security: []
      x-native-ops-role: none
      x-native-ops-enabled-by: --oidc-issuer
      parameters:
        - { name: state, in: query, required: true, schema: { type: string } }
        - { name: code, in: query, required: true, schema: { type: string } }
      responses:
        "302": { description: "Back to the UI, signed in." }
        "403": { description: "The provider's user has no access (with --oidc-role none, until it is granted)." }
  /v1/whoami:
    get:
      operationId: whoami
      tags: [meta]
      summary: Which token and role you are
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      responses:
        "200":
          description: The caller.
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  role: { $ref: "#/components/schemas/Role" }
                  version: { type: string }
                  scope: { $ref: "#/components/schemas/Scope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/status:
    get:
      operationId: status
      tags: [host]
      summary: The host's state
      description: |
        Instances, data volumes, images, host metrics (load, memory, swap, pool disk, uptime) and
        warnings; with the edge and DNS configured, the edge's routes and certificates and the DNS
        records. Instance config is reported by key name only. Cached for a few seconds.
      x-native-ops-role: viewer
      responses:
        "200": { description: "The snapshot.", content: { application/json: { schema: { type: object } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/HostUnavailable" }
  /v1/deploy:
    post:
      operationId: deploy
      tags: [deploys]
      summary: Deploy a protected tag
      description: |
        Starts a deploy job for the commit the tag points at. The job:

        1. resolves the tag on the git server and checks a protection rule covers it (otherwise it fails
           without fetching anything);
        2. downloads that commit's tree from the git server;
        3. if fleet.yml pins another daemon release, upgrades to it, restarts, and resumes the deploy as
           a new job on the new binary (this job ends with result `daemon_upgraded`);
        4. plans it; a blocked plan fails the job and nothing changes; nothing to change ends it with
           result `no_changes`;
        5. applies the plan (result `applied`).

        Follow the job with `GET /v1/jobs/{id}?wait=60`, or see it in `GET /v1/deploys`.
      x-native-ops-role: deployer
      x-native-ops-enabled-by: --enable-apply, --git-url and --deploy-repo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagRequest" }
      responses:
        "202":
          description: The deploy job started. Location names the job.
          headers:
            Location: { schema: { type: string, examples: [/v1/jobs/j-1791054271-18f384d8] } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  job: { $ref: "#/components/schemas/Job" }
                  tag: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Busy" }
  /v1/deploy/plan:
    post:
      operationId: planDeploy
      tags: [deploys]
      summary: What deploying a tag would change
      description: |
        Resolves the tag, reads its commit from the git server and plans it against the host now,
        without changing anything and without starting a job. The plan is recorded (`GET /v1/plans`).

        An unprotected tag is still planned, with `deployable: false` and the reason. When fleet.yml pins
        a daemon release other than the running one, `daemon.upgrade` is true: the deploy would upgrade
        first, and the pinned release makes the plan that is applied.
      x-native-ops-role: planner
      x-native-ops-enabled-by: --enable-apply, --git-url and --deploy-repo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagRequest" }
      responses:
        "200":
          description: The plan.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeployPlan" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: The git server could not resolve the tag (no such tag, or not a deploy tag).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { description: "Too many plans are running; retry after the Retry-After header." }
        "502":
          description: The commit could not be read from the git server (git_unavailable), or the host could not be read (host_unavailable).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/deploys:
    get:
      operationId: listDeploys
      tags: [deploys]
      summary: What the host runs, and the deploy history
      description: |
        `current` is the newest deploy that left the host matching its tag (result `applied` or
        `no_changes`). Changes made since by other means (an apply, an instance change, an image build)
        are in `GET /v1/jobs`. `running` is the deploy in progress, if any.
      x-native-ops-role: viewer
      x-native-ops-enabled-by: --enable-apply, --git-url and --deploy-repo
      parameters:
        - { name: limit, in: query, description: "How many deploys to list (default 50).", schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: The deploys.
          content:
            application/json:
              schema:
                type: object
                required: [current, running, deploys]
                properties:
                  current: { oneOf: [{ $ref: "#/components/schemas/Deploy" }, { type: "null" }] }
                  running: { oneOf: [{ $ref: "#/components/schemas/Deploy" }, { type: "null" }] }
                  deploys: { type: array, items: { $ref: "#/components/schemas/Deploy" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Jobs, newest first (without logs)
      description: A scoped token sees only jobs it started and jobs about instances its scope allows.
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      parameters:
        - name: kind
          in: query
          description: A job kind, or a family of kinds ending in a colon (`instance:`).
          schema: { type: string, examples: [deploy, apply, "instance:", image:build, image:prune, edge:apply, backup, restore, dns:sync, daemon:upgrade] }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/JobStatus" } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: The jobs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs: { type: array, items: { $ref: "#/components/schemas/Job" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: One job
      description: |
        The log is included for deployers and admins only. With `wait`, a running job is answered as soon
        as it ends, or when the wait is over (still running), so a client need not poll.
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, pattern: "^j-[0-9]{9,12}-[0-9a-f]{8}$" } }
        - { name: wait, in: query, description: "Seconds to wait for a running job to end.", schema: { type: integer, minimum: 0, maximum: 60 } }
      responses:
        "200": { description: "The job.", content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/plan:
    post:
      operationId: plan
      tags: [plans]
      summary: Plan an uploaded configuration tree
      description: What apply would change for the uploaded tree (never changes the host). The plan is recorded.
      x-native-ops-role: planner
      parameters:
        - $ref: "#/components/parameters/Service"
        - $ref: "#/components/parameters/Sha"
      requestBody: { $ref: "#/components/requestBodies/Tree" }
      responses:
        "200":
          description: The plan.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sha: { type: string }
                  hash: { type: string }
                  exit: { $ref: "#/components/schemas/PlanExit" }
                  plan: { $ref: "#/components/schemas/FleetPlan" }
                  text: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "415": { description: "The body must be a gzip-compressed tar (Content-Type application/gzip)." }
        "429": { description: "Too many plans are running." }
        "502": { $ref: "#/components/responses/HostUnavailable" }
  /v1/plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: The plans that were made, and where each stands
      x-native-ops-role: viewer
      responses:
        "200": { description: "The plans (pending, approved, used, expired, stale, blocked).", content: { application/json: {} } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: An API token (nops_...), created by an admin (POST /v1/tokens or native-ops token create). The UI's session cookie also works, except at /mcp.
...